mctp i2c: convert tx workqueue to a kthread
In a future change, we'll want to hold the i2c bus locked over a series
of packet tx & rx. A workqueue enforces balanced locking over invocation
of a work function, so is no longer suitable for us.

This change converts the work queue to a kthread, woken when we want to
start tx.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>