drivers: soc: microwatt: Add driver for Microwatt LPC-CTRL
This commit adds a driver for LPC-CTRL peripheral present on the Boxarty
fork of Microwatt project [0]. It is based on another driver for LPC
controller present in Aspeed SoCs (drivers/soc/aspeed/aspeed-lpc-ctrl.c).

It supports loading a firmware to a memory region and makes it avaliable
via LPC FW cycles to an LPC host.

[0] https://github.com/antonblanchard/microwatt/