net: mctp: serial: use netif_receive_skb instead of netif_rx
In the change from netif_rx_ni in b903117b, we ended up converging on
netif_rx for MCTP serial packet ingress. In backlog mode, netif_rx will
end up queueing to a CPU-specific queue, which may result in
out-of-order packets when those queues are processed.

The MCTP spec doesn't have provision for packet reordering, so we need
to be particularly careful to prevent correct sequencing of reassembly
flows.

Use netif_receive_skb() for packet ingress instead; this means we won't
end up with packets on the per-cpu NAPI backlog queue.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>